+++ src/incremental-write.ts
@@ -3,7 +3,7 @@ import {
     rxStorageWriteErrorToRxError
 } from './rx-error';
 import type {
-    AtomicUpdateFunction,
+    ModifyFunction,
     BulkWriteRow,
     MaybePromise,
     RxDocumentData,
@@ -187,7 +187,7 @@ export class IncrementalWriteQueue<RxDocType> {
 
 
 export function incrementalModifierFromPublicToInternal<RxDocType>(
-    publicModifier: AtomicUpdateFunction<RxDocType>
+    publicModifier: ModifyFunction<RxDocType>
 ): IncrementalWriteModifier<RxDocType> {
     const ret = async (docData: RxDocumentData<RxDocType>) => {
         const withoutMeta: WithDeleted<RxDocType> = stripMetaDataFromDocument(docData) as any;