+++ test/unit/rx-document.test.ts
@@ -172,15 +172,15 @@ describe('rx-document.test.js', () => {
                 assert.ok(doc);
 
                 // update some times to generate revisions
-                await doc.atomicUpdate((docData: any) => {
+                await doc.incrementalModify((docData: any) => {
                     docData.age++;
                     return docData;
                 });
-                await doc.atomicUpdate((docData: any) => {
+                await doc.incrementalModify((docData: any) => {
                     docData.age++;
                     return docData;
                 });
-                await doc.atomicUpdate((docData: any) => {
+                await doc.incrementalModify((docData: any) => {
                     docData.age = 100;
                     return docData;
                 });
@@ -209,7 +209,7 @@ describe('rx-document.test.js', () => {
                 assert.ok(docs.length > 1);
                 let first = docs[0];
 
-                first = await first.atomicPatch({ firstName: 'foobar' });
+                first = await first.incrementalPatch({ firstName: 'foobar' });
                 await first.remove();
 
                 const docsAfter = await c.find().exec();
@@ -314,12 +314,12 @@ describe('rx-document.test.js', () => {
             });
         });
     });
-    config.parallel('.atomicUpdate()', () => {
+    config.parallel('.incrementalModify()', () => {
         describe('positive', () => {
             it('run one update', async () => {
                 const c = await humansCollection.createNested(1);
                 let doc = await c.findOne().exec(true);
-                doc = await doc.atomicUpdate((innerDoc: any) => {
+                doc = await doc.incrementalModify((innerDoc: any) => {
                     innerDoc.firstName = 'foobar';
                     return innerDoc;
                 });
@@ -327,7 +327,7 @@ describe('rx-document.test.js', () => {
 
                 /**
                  * Running a totally new query must return the
-                 * exact same document instance as atomicUpdate() did.
+                 * exact same document instance as incrementalModify() did.
                  */
                 const doc2 = await c.findOne().exec(true);
                 assert.ok(doc === doc2);
@@ -338,11 +338,11 @@ describe('rx-document.test.js', () => {
                 const c = await humansCollection.createNested(1);
                 let doc = await c.findOne().exec(true);
 
-                doc.atomicUpdate((innerDoc: any) => {
+                doc.incrementalModify((innerDoc: any) => {
                     innerDoc.firstName = 'foobar';
                     return innerDoc;
                 });
-                await doc.atomicUpdate((innerDoc: any) => {
+                await doc.incrementalModify((innerDoc: any) => {
                     innerDoc.firstName = 'foobar2';
                     return innerDoc;
                 });
@@ -361,7 +361,7 @@ describe('rx-document.test.js', () => {
                         t++;
                         return t;
                     })
-                    .forEach(x => lastPromise = doc.atomicUpdate((innerDoc: any) => {
+                    .forEach(x => lastPromise = doc.incrementalModify((innerDoc: any) => {
                         innerDoc.age = x;
                         return innerDoc;
                     }));
@@ -381,7 +381,7 @@ describe('rx-document.test.js', () => {
                         t++;
                         return t;
                     })
-                    .forEach(x => lastPromise = doc.atomicUpdate(async (innerDoc: any) => {
+                    .forEach(x => lastPromise = doc.incrementalModify(async (innerDoc: any) => {
                         await promiseWait(1);
                         innerDoc.age = x;
                         return innerDoc;
@@ -411,19 +411,19 @@ describe('rx-document.test.js', () => {
                 await c.insert(schemaObjects.simpleHuman());
                 const doc = await c.findOne().exec();
 
-                doc.atomicUpdate((innerDoc: any) => {
+                doc.incrementalModify((innerDoc: any) => {
                     innerDoc.firstName = 'foobar';
                     return innerDoc;
                 });
 
-                await doc.atomicUpdate((innerDoc: any) => {
+                await doc.incrementalModify((innerDoc: any) => {
                     innerDoc.firstName = 'foobar2';
                     return innerDoc;
                 });
 
                 await AsyncTestUtil.wait(50);
 
-                await doc.atomicUpdate((innerDoc: any) => {
+                await doc.incrementalModify((innerDoc: any) => {
                     innerDoc.firstName = 'foobar3';
                     return innerDoc;
                 });
@@ -451,7 +451,7 @@ describe('rx-document.test.js', () => {
                 let doc = await c.findOne().exec(true);
                 const docData = doc.toJSON();
                 assert.ok(docData);
-                doc = await doc.atomicUpdate((innerDoc: any) => {
+                doc = await doc.incrementalModify((innerDoc: any) => {
                     innerDoc.firstName = 'foobar';
                     return innerDoc;
                 });
@@ -509,11 +509,11 @@ describe('rx-document.test.js', () => {
                 const doc2 = await c2.findOne().exec(true);
 
                 await Promise.all([
-                    doc.atomicUpdate((d: any) => {
+                    doc.incrementalModify((d: any) => {
                         d.firstName = 'foobar1';
                         return d;
                     }),
-                    doc2.atomicUpdate((d: any) => {
+                    doc2.incrementalModify((d: any) => {
                         d.firstName = 'foobar2';
                         return d;
                     })
@@ -545,7 +545,7 @@ describe('rx-document.test.js', () => {
                 const doc = await col.insert(docData);
 
                 await AsyncTestUtil.assertThrows(
-                    () => doc.atomicUpdate((data: any) => {
+                    () => doc.incrementalModify((data: any) => {
                         data.age = 100;
                         return data;
                     }),
@@ -560,21 +560,21 @@ describe('rx-document.test.js', () => {
 
                 // non-async mutation
                 try {
-                    await doc.atomicUpdate(() => {
+                    await doc.incrementalModify(() => {
                         throw new Error('throws intentional A');
                     });
                 } catch (err) { }
 
                 // async mutation
                 try {
-                    await doc.atomicUpdate(async () => {
+                    await doc.incrementalModify(async () => {
                         await wait(10);
                         throw new Error('throws intentional B');
                     });
                 } catch (err) { }
 
                 // non throwing mutation
-                doc = await doc.atomicUpdate(d => {
+                doc = await doc.incrementalModify(d => {
                     d.age = 150;
                     return d;
                 });
@@ -584,13 +584,13 @@ describe('rx-document.test.js', () => {
             });
         });
     });
-    config.parallel('.atomicPatch()', () => {
+    config.parallel('.incrementalPatch()', () => {
         describe('positive', () => {
             it('run one update', async () => {
                 const c = await humansCollection.createNested(1);
                 const doc = await c.findOne().exec(true);
 
-                const returnedDoc = await doc.atomicPatch({
+                const returnedDoc = await doc.incrementalPatch({
                     firstName: 'foobar'
                 });
                 assert.strictEqual('foobar', returnedDoc.firstName);
@@ -605,7 +605,7 @@ describe('rx-document.test.js', () => {
 
                 assert.ok(doc.mainSkill);
 
-                doc = await doc.atomicPatch({
+                doc = await doc.incrementalPatch({
                     mainSkill: undefined
                 });
 
@@ -737,7 +737,7 @@ describe('rx-document.test.js', () => {
                     value = newVal;
                 });
 
-                await doc.atomicPatch({ firstName: 'foobar' });
+                await doc.incrementalPatch({ firstName: 'foobar' });
 
                 await promiseWait(5);
                 assert.strictEqual(value, 'foobar');
@@ -762,7 +762,7 @@ describe('rx-document.test.js', () => {
                     value = newVal;
                 });
 
-                await doc.atomicPatch({
+                await doc.incrementalPatch({
                     mainSkill: {
                         name: randomCouchString(5),
                         level: 10
@@ -782,7 +782,7 @@ describe('rx-document.test.js', () => {
                 (doc.mainSkill.attack as any).good$.subscribe((newVal: any) => {
                     value = newVal;
                 });
-                await doc.atomicPatch({
+                await doc.incrementalPatch({
                     mainSkill: {
                         name: 'foobar',
                         attack: {
@@ -869,7 +869,7 @@ describe('rx-document.test.js', () => {
             assert.strictEqual(doc.skills.length, 3);
 
             const newSkill = 'newSikSkill';
-            await doc.atomicPatch({ skills: doc.skills.concat(newSkill) });
+            await doc.incrementalPatch({ skills: doc.skills.concat(newSkill) });
 
             const colDump = await col.exportJSON();
             const afterSkills = colDump.docs[0].skills;